agentex/pyproject.toml
[project]
name = "agentex-backend"
version = "0.1.0"
description = ""
authors = [{ name = "Felix Su", email = "felix.su@scale.com" }]
requires-python = ">=3.12,<3.13"
readme = "README.md"
dependencies = [
    "fastapi>=0.115.0,<0.116",
    "litellm>=1.48.2,<2",
    "python-dotenv>=1.0.1,<2",
    "temporalio>=1.18.0,<2",
    "uvicorn[standard]>=0.35.0,<0.36",
    "redis>=5.1.0,<6",
    "sqlalchemy>=2.0.35,<3",
    "asyncpg>=0.29.0,<0.30",
    "alembic>=1.13.3,<2",
    "psycopg2-binary>=2.9.9,<3",
    "docker>=7.1.0,<8",
    "python-multipart>=0.0.12,<0.0.13",
    "aiodocker>=0.23.0,<0.24",
    "kubernetes-asyncio>=31.1.0,<32",
    "aiohttp>=3.10.9,<4",
    "websockets~=14.2",
    "pymongo>=4.11.2,<5",
    "httpx[http2]>=0.27.2",
    "ddtrace>=3.13.0",
    "json_log_formatter>=1.1.1",
    "datadog>=0.52.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.3.3,<9",
    "pytest-asyncio>=1.0.0,<2",
    "agentex-sdk",
    "pytest-cov>=5.0.0,<6",
    "vulture>=2.14",
    "ruff>=0.3.4",
]
test = [
    "pytest>=8.3.3,<9",
    "pytest-asyncio>=1.0.0,<2",
    "pytest-cov>=5.0.0,<6",
    "testcontainers>=4.0.0,<5",
    "httpx[http2]>=0.27.0,<0.28", # for test client
    "factory-boy>=3.3.0,<4", # for test data factories
    "greenlet>=3.2.3",
    "asyncpg>=0.29.0",
]
docs = [
    "mkdocs-material>=9.6.14,<10",
    "mkdocs-macros-plugin>=1.3.7,<2",
    "mkdocstrings-python>=1.16.12",
    "mkdocs>=1.6.1",
    "agentex-sdk",
    "griffe-pydantic>=1.1.4",
    "pymdown-extensions>=10.0,<11",
    "agentex-sdk",
]

[tool.hatch.build.targets.sdist]
include = ["src"]

[tool.hatch.build.targets.wheel]
include = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
